feat: CollapseMessageError styled file
